refactor: table components and styles
